Support arguments in sub field selection in DGS Client query APIs
